+++ src/infrastructure/router/rider/rider.router.ts
@@ -80,7 +80,7 @@ const schema = buildSchema(`
   
   type Call {
     id: String
-    callerRatingStars: Int!
+    callerRatingStars: Float!
     requestedVehicles: RequestedVehicles!
     priceInCents: Int!
     description: String!
